feat: implements new() and bind_services() initial versions
`new()` returns an initialized instance of NetworkListener and will be
called by the `Controller`

`bind_services()` populates the `listeners` and instanciates the
`service_detector`. It will also be called by the `Controller`

**Some changes will be needed in the UML:**
- `new()`: Now takes the `session_tx` as a param. Controller should
  initialize the channel before getting a `NetworkListener` instance
- `start_listening()`: In reflection to the previous change, it doesn't
  take the `session_tx` as a parameter

*Additional comments:*
Files other than `network_listener.rs` have been modified with only what
was needed for the network_listener to work, this is not the actual
implementation of these sub-modules. Nonetheless some of the changes are
required to stay in the merged version