feat: implement `session_manager` module (#96)
* fix: remove punctuation in hello world (#58)

* feat: Add skeleton of each struct

* feat: Implement all functions except one

* feat: implement handle_session and setup_data_proxy

* feat: implement NetworkListener (#75)

* feat: initial commit

Setup `NetworkListener` structure adding minimal code needed for the
types and fields

Import of tokio crate

* feat: initial commit

Setup `NetworkListener` structure adding minimal code needed for the
types and fields

Import of tokio crate

* feat: implements new() and bind_services() initial versions

`new()` returns an initialized instance of NetworkListener and will be
called by the `Controller`

`bind_services()` populates the `listeners` and instanciates the
`service_detector`. It will also be called by the `Controller`

**Some changes will be needed in the UML:**
- `new()`: Now takes the `session_tx` as a param. Controller should
  initialize the channel before getting a `NetworkListener` instance
- `start_listening()`: In reflection to the previous change, it doesn't
  take the `session_tx` as a parameter

*Additional comments:*
Files other than `network_listener.rs` have been modified with only what
was needed for the network_listener to work, this is not the actual
implementation of these sub-modules. Nonetheless some of the changes are
required to stay in the merged version

* test: bind_services

Unit tests now passes for the bind_services with a legit
ServiceConfiguration

* feat: implements start_listening function

Not yet done, but function capable of sending a SessionRequest through a
mpsc channel and be read from the other side

Because port must be actually opened for the TcpStream to be able to
connect in the SessionRequest we needed to have an actual service that
opens a port

`start_http_server()` spawns a thread listening on port 8000 on
localhost so TcpStream can connect to it to workaround for the moment

* doc: document NetworkListener sub-module

Documentation done for:
- Global sub-module is documented with examples and a sequence diagram
- The structure is documented with every field being described
- Every function has documentation with return type, error outputs and examples

* feat: implement whole sequence of NetworkListener

Starts with bind_services that creates a new ServiceDetector and fills the listeners with a port per service

Then start_listening binds the sockets with an address and transfers it to the listen_on_port which waits for an incoming connection, accepts it and transfers the stream and client_addr to the handle_connection which is responsible for detecting the service requested, formulate the SessionRequest and send it to the session manager.

Connection is filtered in the listen_on_port when a new connection is incoming with the help of the ConnectionFilter

* test: tests methods from NetworkListener

start_listening, listen_on_port and handle_connection each have unit
tests that pass

* feat: implements ConnectionFilter

Implements methods from the ConnectionFilter

is_connection_allowed() is called from NetworkListener and uses
is_ip_allowed and is_port_allowed internally

is_ip_allowed decides depending on the whitelist_default field

is_port_allowed is blacklist by default

* doc: update class diagram

Update for NetworkListener and ConnectionFilter

* feat: delete ProtocolFilter

Useless because not in requirements

* feat: implement ServiceDetector

* bugfix: debug and cleanup

Debug of listen_on_port test that would panic because nothing was
contained in the stream. Now specifies a port in the ServiceConfig
passed to the ServiceDetector so service detection by port works

Clean by removing commented functions not in use and useless tests

Still remains some points to solve concerning addressing and binding
(everything is in comments in the network_listener.rs)

* fix: fix linting issue

* fix: ignore example code in rustdoc

* fix: replace non-blocking try_read with blocking read

* fix: update new() in NetworkListener

NetworkListener now gets the IP it should bind the listeners to through
the new() function

---------

Co-authored-by: nilsdonatantonio <nils.donatantonio@heig-vd.ch>

* feat: Add skeleton of each struct

* feat: Implement all functions except one

* fix: add mock storage to build session_manager

* fix: add mock storage to build session_manager

* feat: Correct several syntax errors

* feat: adjust handle_session

* fix: handle_session and setup_data_proxy

* fix: cleanup_expired_sessions (use retain to remove)

* fix: cleanup_expired_sessions, shutdown_all_sessions and create_session

* feat: implement join_streams and find_session function

* refactor: rework create_session and handle_session

* chore: clean up rebase

* fix: max_session check and use retain to cleanup expired sessions

* feat: refactor session management and update session structure

* doc: enhance ActiveSession documentation

* fix: update session import to use new session module

* feat: update service ports and web UI port configuration

* feat: update web UI port configuration to 3000

* fix: solve type issues throughout the program

* refactor: update storage demo for clarity

* chore: rename file

* feat: make Controller::new async and update storage initialization

* feat: enhance logging configuration

* chore: add miel.sqlite3 to .gitignore

* fix: refactor session management handle database interactions correctly

* fix: support asynchronous code and update database initialization

* feat: integrate storage components into SessionManager

* style: format source code

* fix: update code example to use `ignore` type

---------

Co-authored-by: EstebanGiorgis <145534788+EstebanGiorgis@users.noreply.github.com>
Co-authored-by: Drin Racaj <drin.racaj@hes-so.ch>
Co-authored-by: Drin Racaj <145534814+drinracaj@users.noreply.github.com>
Co-authored-by: nilsdonatantonio <nils.donatantonio@heig-vd.ch>